Refactor FaktaGrunnlag-filer for å kun ha FaktaGrunnlagData-filer i sub-paket
