+++ src/main/kotlin/no/nav/tilleggsstonader/sak/opplysninger/grunnlag/FaktaGrunnlag.kt
@@ -1,19 +1,20 @@
-package no.nav.tilleggsstonader.sak.opplysninger.grunnlag.faktagrunnlag
+package no.nav.tilleggsstonader.sak.opplysninger.grunnlag
 
 import no.nav.tilleggsstonader.sak.felles.domain.BehandlingId
 import no.nav.tilleggsstonader.sak.felles.domain.FaktaGrunnlagId
 import no.nav.tilleggsstonader.sak.infrastruktur.database.Sporbar
 import no.nav.tilleggsstonader.sak.infrastruktur.exception.feilHvis
+import no.nav.tilleggsstonader.sak.opplysninger.grunnlag.faktagrunnlag.FaktaGrunnlagData
+import no.nav.tilleggsstonader.sak.opplysninger.grunnlag.faktagrunnlag.TypeFaktaGrunnlag
 import org.springframework.data.annotation.Id
 import org.springframework.data.relational.core.mapping.Embedded
 import org.springframework.data.relational.core.mapping.Table
-import kotlin.reflect.KClass
 
 typealias FaktaGrunnlag = GeneriskFaktaGrunnlag<out FaktaGrunnlagData>
 
 /**
  * [behandlingId], [type] og [typeId] er unik sammen
- * @param typeId er en ekstra identifikator, eks ident på barnet for [TypeFaktaGrunnlag.BARN_ANDRE_FORELDRE_SAKSINFORMASJON]
+ * @param typeId er en ekstra identifikator, eks ident på barnet for [no.nav.tilleggsstonader.sak.opplysninger.grunnlag.faktagrunnlag.TypeFaktaGrunnlag.BARN_ANDRE_FORELDRE_SAKSINFORMASJON]
  */
 @Table("fakta_grunnlag")
 data class GeneriskFaktaGrunnlag<T : FaktaGrunnlagData>(
@@ -30,24 +31,5 @@ data class GeneriskFaktaGrunnlag<T : FaktaGrunnlagData>(
         feilHvis(data.type != type) {
             "Type=$type er ikke lik type i data (${data.type})"
         }
-    }
-}
-
-sealed interface FaktaGrunnlagData : FaktaGrunnlagDataJson {
-    val type: TypeFaktaGrunnlag
-}
-
-enum class TypeFaktaGrunnlag(
-    val kClass: KClass<out FaktaGrunnlagData>,
-) {
-    PERSONOPPLYSNINGER(FaktaGrunnlagPersonopplysninger::class),
-    BARN_ANDRE_FORELDRE_SAKSINFORMASJON(FaktaGrunnlagBarnAndreForeldreSaksinformasjon::class),
-    ARENA_VEDTAK_TOM(FaktaGrunnlagArenaVedtak::class),
-    ;
-
-    companion object {
-        private val perType = entries.associateBy(TypeFaktaGrunnlag::kClass)
-
-        fun finnType(type: KClass<out FaktaGrunnlagData>) = perType[type] ?: error("Finner ikke type for ${type.simpleName}")
     }
 }
