+++ src/main/kotlin/no/nav/tilleggsstonader/sak/opplysninger/grunnlag/Grunnlag.kt
@@ -3,7 +3,6 @@ package no.nav.tilleggsstonader.sak.opplysninger.grunnlag
 import no.nav.tilleggsstonader.sak.opplysninger.grunnlag.faktagrunnlag.FaktaGrunnlagArenaVedtak
 import no.nav.tilleggsstonader.sak.opplysninger.grunnlag.faktagrunnlag.FaktaGrunnlagBarnAndreForeldreSaksinformasjon
 import no.nav.tilleggsstonader.sak.opplysninger.grunnlag.faktagrunnlag.FaktaGrunnlagPersonopplysninger
-import no.nav.tilleggsstonader.sak.opplysninger.grunnlag.faktagrunnlag.GeneriskFaktaGrunnlag
 
 data class Grunnlag(
     val personopplysninger: FaktaGrunnlagPersonopplysninger,
