+++ src/test/kotlin/no/nav/tilleggsstonader/sak/infrastruktur/OpprettGrunnlagsdataLokalPostgres.kt
@@ -3,7 +3,7 @@ package no.nav.tilleggsstonader.sak.infrastruktur
 import no.nav.tilleggsstonader.libs.log.SecureLogger.secureLogger
 import no.nav.tilleggsstonader.sak.behandling.domain.BehandlingRepository
 import no.nav.tilleggsstonader.sak.behandling.domain.BehandlingStatus
-import no.nav.tilleggsstonader.sak.opplysninger.grunnlag.faktagrunnlag.FaktaGrunnlagService
+import no.nav.tilleggsstonader.sak.opplysninger.grunnlag.FaktaGrunnlagService
 import org.springframework.context.annotation.Configuration
 import org.springframework.context.annotation.Profile
