+++ src/test/kotlin/no/nav/tilleggsstonader/sak/opplysninger/grunnlag/FaktaGrunnlagRepositoryTest.kt
@@ -1,9 +1,10 @@
-package no.nav.tilleggsstonader.sak.opplysninger.grunnlag.faktagrunnlag
+package no.nav.tilleggsstonader.sak.opplysninger.grunnlag
 
 import no.nav.tilleggsstonader.sak.IntegrationTest
-import no.nav.tilleggsstonader.sak.opplysninger.grunnlag.faktagrunnlag.GeneriskFaktaGrunnlagTestUtil.faktaGrunnlagBarnAnnenForelder
+import no.nav.tilleggsstonader.sak.opplysninger.grunnlag.faktagrunnlag.GeneriskFaktaGrunnlagTestUtil
+import no.nav.tilleggsstonader.sak.opplysninger.grunnlag.faktagrunnlag.TypeFaktaGrunnlag
 import no.nav.tilleggsstonader.sak.util.behandling
-import org.assertj.core.api.Assertions.assertThat
+import org.assertj.core.api.Assertions
 import org.junit.jupiter.api.Test
 import org.springframework.beans.factory.annotation.Autowired
 
@@ -15,7 +16,8 @@ class FaktaGrunnlagRepositoryTest : IntegrationTest() {
     fun `skal kunne lagre og hente opp data`() {
         val behandling = testoppsettService.opprettBehandlingMedFagsak(behandling())
 
-        val faktaGrunnlagBarnAnnenForelder = faktaGrunnlagBarnAnnenForelder(behandlingId = behandling.id)
+        val faktaGrunnlagBarnAnnenForelder =
+            GeneriskFaktaGrunnlagTestUtil.faktaGrunnlagBarnAnnenForelder(behandlingId = behandling.id)
         faktaGrunnlagRepository.insert(faktaGrunnlagBarnAnnenForelder)
 
         val fakta =
@@ -23,9 +25,9 @@ class FaktaGrunnlagRepositoryTest : IntegrationTest() {
                 behandlingId = behandling.id,
                 type = TypeFaktaGrunnlag.BARN_ANDRE_FORELDRE_SAKSINFORMASJON,
             )
-        assertThat(fakta).hasSize(1)
-        assertThat(fakta[0].data).isEqualTo(faktaGrunnlagBarnAnnenForelder.data)
-        assertThat(fakta[0].typeId).isEqualTo("barn1")
-        assertThat(fakta[0].type).isEqualTo(TypeFaktaGrunnlag.BARN_ANDRE_FORELDRE_SAKSINFORMASJON)
+        Assertions.assertThat(fakta).hasSize(1)
+        Assertions.assertThat(fakta[0].data).isEqualTo(faktaGrunnlagBarnAnnenForelder.data)
+        Assertions.assertThat(fakta[0].typeId).isEqualTo("barn1")
+        Assertions.assertThat(fakta[0].type).isEqualTo(TypeFaktaGrunnlag.BARN_ANDRE_FORELDRE_SAKSINFORMASJON)
     }
 }
